+++ app/src/Controller/UserController.php
@@ -32,6 +32,8 @@ class UserController extends AbstractController
     public function new(Request $request, UserPasswordHasherInterface $passwordEncoder): Response
     {
         $user = new User();
+        $this->denyAccessUnlessGranted('CREATE', $user);
+
         $form = $this->createForm(UserType::class, $user);
         $form->handleRequest($request);
 
@@ -56,6 +58,8 @@ class UserController extends AbstractController
      */
     public function show(User $user): Response
     {
+        $this->denyAccessUnlessGranted('SHOW', $user);
+
         return $this->render('user/show.html.twig', [
             'user' => $user,
         ]);
@@ -66,6 +70,8 @@ class UserController extends AbstractController
      */
     public function edit(Request $request, User $user): Response
     {
+        $this->denyAccessUnlessGranted('EDIT', $user);
+
         $form = $this->createForm(UserType::class, $user);
         $form->handleRequest($request);
 
@@ -86,6 +92,8 @@ class UserController extends AbstractController
      */
     public function delete(Request $request, User $user): Response
     {
+        $this->denyAccessUnlessGranted('DELETE', $user);
+
         if ($this->isCsrfTokenValid('delete'.$user->getId(), $request->request->get('_token'))) {
             $entityManager = $this->getDoctrine()->getManager();
             $entityManager->remove($user);
